pyproject.toml
[tool.poetry]
name = "zit test"
version = "0.1.0"
description = ""
authors = ["barabane <barabayne@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.34"
fastapi = "^0.114.2"
pydantic-settings = "^2.5.2"
pre-commit = "^3.8.0"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
uvicorn = "^0.30.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
